Added a lot of layout elements to the content of scroll view and buttons
Based on the official tutorials from the Unity documentation. Link: https://unity3d.com/learn/tutorials/topics/user-interface-ui/scroll-view?playlist=17111
Added vertical layout group to content for arranging the buttons.
Added content size fitter to the content for making the buttons fit the width but we unchecked enforcing the height. We set using minimum height for that.
Addd Layout element to button for specifying that what the minimum height of a will be.