FOUND WORK AROUND TO FUCKING UNITY EDITOR BUT >:(
Had just finished with the scripts for dynamically displaying a list of items from the object pool based on the prefab template.

For some reason, Unity messed up the scale values (transform.localscale x y z) of each prefab or instance of the object. Had to manually readjust them to 1 in the code.

I'm sure it was the Unity editor because when dragging the prefabs manually in the editor to the hierarchy it worked perfectly. Only when instantiating them dynamically the scale changed to a werid value of 2.something...

USEFUL DOCUMENTATION THAT GAVE ME THE IDEA FOR THE WORKAROUND:
- Scaling dynamic created buttons (https://answers.unity.com/questions/1257149/scaling-dynamic-created-buttons.html)

Based on the Unity tutorials. Link: https://unity3d.com/learn/tutorials/topics/user-interface-ui/creating-items-and-testing?playlist=17111